+++ lib/core/utils/session_serializer.dart
@@ -1,0 +1,4 @@
+abstract class SessionSerializer<T> {
+  T? fromJson(dynamic json);
+  Map<String, dynamic> toJson(T session);
+}