+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 ## 1.0.1 - 2026-1-13
 
 - Session data type now can be any type
+
+## 1.0.2 - 2026-1-13
+
+- Fix `SessionData` generic type error
+- Added `session_serializer` interface to serialize and deserialize `SessionData` to and from JSON to pass to the server